[MOV] mail: move unused chatter props from portal
Move chatter props which are not used in portal and their related methods from
portal to web folder (two props `webRecord` and `saveRecord` will be moved in
a followup PR)
